solved problem
